% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ask}
\alias{ask}
\title{Ask the Mystic Seer a question}
\usage{
ask(question, x)
}
\arguments{
\item{question}{a question (character string) to which an answer is sought}

\item{x}{an object pertaining to the question}
}
\description{
The Mystic Seer will ponder your question before giving an answer; Best when used in the Twilight Zone.
}
