% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ask}
\alias{ask}
\title{Ask the Mystic Seer a question}
\usage{
ask(question, x)
}
\arguments{
\item{question}{a question (character string)}

\item{x}{an object pertaining to the question}
}
\value{
the answer (character string)
}
\description{
The Mystic Seer produces an answer to your Yes/No question immediately.
}
\examples{
ask("Should I go to New York City for my honeymoon?")
ask("Did I get that promotion?")

ask("Does this dataset contain all the secrets I need to know?", iris)

cars_model <- lm(dist ~ speed, data = cars)
ask("Must I tweak this model some more?", cars_model)
}
